# Removing unnecessary iterator operations from op_apply