In many cases we are iterating tensors with one continuous memory (any stride size). That means there is no need to check overflow/increase counters. It saves 3 asm operations per tensor (out of 13 in copy case).

Simple test:

```
import torch
torch.set_num_threads(1)
i = torch.arange(100000000).type(torch.DoubleTensor).reshape(20,500,100,100)
timeit i.transpose(0,2).transpose(1,2).clone()
```

With optimization:
962 ms ± 17.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

Without:
1.03 s ± 22.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

There is still room to apply it in all CPU_tensor_* functions. I just covered most common ones.