@ezyang 2019-02-27
I found this diff somewhat hard to understand, I think part of it is due to some funny business in the current code structure. It might be right to take this patch as is, but I think this patch could be clearer if we did a little refactoring first.

Right now, we first test what the dimensionality of the tensors are (to decide if we shunt into the optimized codepath) and the collapse the dimensions. This seems wrong: we should *first* collapse dimensions and then decide if we can go in the codepath. There are multiple benefits of this:

1. More tensors will be eligible for the optimized codepath. If I have a 100-dim contiguous tensor, I don't get to use the fixed iterator. There's no reason why I shouldn't be able to use it in this case.
2. The condition for `CanOverflow` is simplified. Now it is equivalent to `N = 1`. So we can eliminate the extra parameter entirely. (You could say `N = 8` in that case, but that's just wasting runtime space.)

One extra thing to note: with this change, we generate an *extra* copy of the kernel, specifically specialized for the contiguous case. This seems reasonable to me, but it's something to be aware of from a binary size perspective. Might be worth measuring what binary size change was.

@cpuhrsch, could you also review this diff (and my comment?)