@VitalyFedyunin 2019-02-27
Btw, I've removed dimensionality test to make sure that we are not slowing down existing functionality. Totally agree about 100dim contiguous tensor thou. I will handle such cases in next PR on top of it. 

Numbers of kernels is a valid concern if we plan to apply same patter to all CPU_ops. As for now it is only 2 additional kernels per op to cover the most popular use-case (clone() function).