@ezyang 2019-02-27
@fmassa wanted to know if we should just use TensorIterator directly instead.