@cpuhrsch 2019-02-27
In general, TensorIterator has supplanted a lot of the code in this file. @colesbury has the necessary context here.

We should get a list of callsites of these functions to see what is affected and whether it's possible replacing it with TensorIterator.

The difference in timing here is pretty small, so it's hard to say how precise it is. It's also restricted to a single function. When building this code we looked at various tensor shapes and in particular unary operations because those were the first customers. This requires precise control of the environment. Vitaly and I are talking about this right now.

cc @apaszke who worked on this with me.