@cpuhrsch 2019-02-27
As in aside, it might be worth looking into removing CPUApplyUtils.h entirely if it's been replaced by TensorIterator. Then this PR's insight can be used to improve TensorIterator and we further reduce the complexity of our codebase.